% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_app.R
\name{open_app}
\alias{open_app}
\title{Open Shiny app for data visualization}
\usage{
open_app(language = c("eng"))
}
\arguments{
\item{language}{select which language the app will have. Options are "pt" for Portuguese and "en" for English.}
}
\value{
None. This function launches the Shiny application.
}
\description{
This function opens a shiny app that uses the datasets provided
}
\details{
This function opens the Shiny application corresponding to the specified language for the Brazilian lottery data.
}
