% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{#' Function to download lottery data from git repository}
\usage{
get_data(
  game = c("maismilionaria", "megasena", "lotofacil", "quina", "lotomania", "timemania",
    "duplasena", "federal", "loteca", "diadesorte", "supersete"),
  type = c("dezenas", "ganhadores")
)
}
\arguments{
\item{game}{select which Brazilian lottery game}

\item{type}{select if you want the winners database or the results}
}
\value{
tibble/data.frame containing the downloaded data
}
\description{
This function downloads the Brazilian lottery game data from Caixa Federal up to the date declared
}
\examples{
\dontrun{
library(LotteryBrasil)
megasena_ganhadores <- get_data(game= "megasena",type = "ganhadores")
megasena_ganhadores

lotofacil_dezenas <- get_data(game= "lotofacil",type = "dezenas")
lotofacil_dezenas
}

}
